+++ build_version.hpp
@@ -1,0 +1,5 @@
+﻿#pragma once
+
+#ifndef APP_VERSION_W
+#define APP_VERSION_W L"v0.10.0"
+#endif